+++ Package.swift
@@ -5,6 +5,6 @@ import PackageDescription
 let package = Package(
     name: "LaAccelerate",
     dependencies: [
-        .Package(url: "https://github.com/mitsuse/la", majorVersion: 0, minor: 6),
+        .Package(url: "https://github.com/mitsuse/la", majorVersion: 0, minor: 7),
     ]
 )